// src/ui/themes/types.ts
export type Theme = {
  radius?: string
  colors?: {
    border?: string
    input?: string
    ring?: string
    background?: string
    foreground?: string
    radius?: string
    primary?: {
      DEFAULT?: string
      foreground?: string
    }
    secondary?: {
      DEFAULT?: string
      foreground?: string
    }
    destructive?: {
      DEFAULT?: string
      foreground?: string
    }
    muted?: {
      DEFAULT?: string
      foreground?: string
    }
    accent?: {
      DEFAULT?: string
      foreground?: string
    }
    popover?: {
      DEFAULT?: string
      foreground?: string
    }
    card?: {
      DEFAULT?: string
      foreground?: string
    }
  }
}